silence another warning

+++ src/dft_ldos.cpp
@@ -133,7 +133,7 @@ void dft_ldos::update(fields &f) {
         }
       }
     }
-  for (int i = 0; i < freq.size(); ++i) {
+  for (size_t i = 0; i < freq.size(); ++i) {
     complex<double> Ephase = polar(1.0, 2*pi * freq[i] * f.time()) * scale;
     complex<double> Hphase = polar(1.0, 2*pi * freq[i] * (f.time() - f.dt / 2)) * scale;
     Fdft[i] += Ephase * EJ + Hphase * HJ;
